add compile unit for String.hpp
one less <sstream> in header
