+++ src/opm/input/eclipse/Units/UnitSystem.cpp
@@ -23,6 +23,7 @@
 #include <opm/input/eclipse/Units/Dimension.hpp>
 #include <opm/input/eclipse/Units/Units.hpp>
 
+#include <algorithm>
 #include <cstddef>
 #include <limits>
 #include <stdexcept>
